FIX: Omission in previous change

+++ dataobj/settings.h
@@ -969,13 +969,13 @@ public:
 	bool get_with_private_paks() const { return with_private_paks; }
 
 	// @author: jamespetts
-	uint16 get_min_visiting_tolerance() const { return min_visiting_tolerance; }
+	uint32 get_min_visiting_tolerance() const { return min_visiting_tolerance; }
 	void set_min_visiting_tolerance(uint16 value) { min_visiting_tolerance = value; }
-	uint16 get_range_commuting_tolerance() const { return range_commuting_tolerance; }
+	uint32 get_range_commuting_tolerance() const { return range_commuting_tolerance; }
 	void set_range_commuting_tolerance(uint16 value) { range_commuting_tolerance = value; }
-	uint16 get_min_commuting_tolerance() const { return min_commuting_tolerance; }
+	uint32 get_min_commuting_tolerance() const { return min_commuting_tolerance; }
 	void set_min_commuting_tolerance(uint16 value) { min_commuting_tolerance = value; }
-	uint16 get_range_visiting_tolerance() const { return range_visiting_tolerance; }
+	uint32 get_range_visiting_tolerance() const { return range_visiting_tolerance; }
 	void set_range_visiting_tolerance(uint16 value) { range_visiting_tolerance = value; }
 
 	// town growth stuff
